backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
pyyaml==6.0.1
pydantic==2.5.3
scipy==1.11.4
scikit-image==0.22.0
